chore: Add tests to the parse functions and split
Tests have been added to test the sanity of what is expected vs what is
gotten.
The tests are not comprehensive enough but they cover most use case.

The functions have been split to enable modular testing.
This means some resources may be *mismanaged or leak.
But future mitigations will mitigate this.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>